feat(web): Files as a side panel on the session page (review feedback)
