+++ web/src/App.tsx
@@ -9,7 +9,6 @@ import { ExtensionDialog, type UiRequest } from "./components/ExtensionDialog";
 import { SessionsPage } from "./components/SessionsPage";
 import { AgentPage } from "./components/AgentPage";
 import { RoutinesPage } from "./components/RoutinesPage";
-import { FilesPage } from "./components/FilesPage";
 import { ThemeSwitcher } from "./components/ThemeSwitcher";
 
 // Legacy routes ("session", "global") still resolve — old links stay valid.
@@ -50,7 +49,7 @@ export default function App() {
       <Route path="/sessions" element={<Shell view="sessions" />} />
       <Route path="/agent" element={<Shell view="agent" />} />
       <Route path="/routines" element={<Shell view="routines" />} />
-      <Route path="/files" element={<Shell view="files" />} />
+
       <Route path="/s/:sessionId" element={<Shell />} />
       <Route path="/s/:sessionId/settings" element={<Shell settings />} />
       <Route path="/s/:sessionId/settings/:tab" element={<Shell settings />} />
@@ -66,7 +65,7 @@ function Shell({
   view = "chat",
 }: {
   settings?: boolean;
-  view?: "chat" | "sessions" | "agent" | "routines" | "files";
+  view?: "chat" | "sessions" | "agent" | "routines";
 }) {
   const { sessionId, tab } = useParams<{ sessionId?: string; tab?: string }>();
   const navigate = useNavigate();
@@ -228,8 +227,6 @@ function Shell({
           <AgentPage onSelect={(id) => navigate(`/s/${id}`)} />
         ) : view === "routines" ? (
           <RoutinesPage onOpenSession={(id) => navigate(`/s/${id}`)} />
-        ) : view === "files" ? (
-          <FilesPage workspaces={workspaces} />
         ) : active ? (
           <Chat
             session={active}
